feat: Add functionality for depositing amount
This commit introduces functionality for depositing an amount to a
user's account. It includes the implementation of menu option 2 to allow
users to deposit money. Upon selection this option, users are prompted
to enter their first name, and if found, their account balance is
updated accordingly. The updated balance is then stored in the
Reg_users1.dat file.